Moves util string methods into LoggingHelper class
